Struct ccthw::application::GlfwWindow

pub struct GlfwWindow { /* private fields */ }

All resources required for running a single-windowed GLFW application which renders graphics using Vulkan.

GlfwWindow derefs as a raw GLFW window handle so application state can configure the window however is convenient.

Implementations

impl GlfwWindow

pub fn new(window_title: impl AsRef<str>) -> Result<Self>

Create a new GLFW window.

The window starts in “windowed” mode and can be toggled into fullscreen or resized by the application.

Params

window_title - The title shown on the window’s top bar.

pub fn toggle_fullscreen(&mut self) -> Result<()>

Toggle application fullscreen.

If the window is currently windowed then swap to fullscreen using whatever the primary monitor advertises as the primary video mode.

If the window is currently fullscreen, then swap to windowed and restore the window’s previous size and location.

pub unsafe fn create_default_render_device( &self, physical_device_features: PhysicalDeviceFeatures ) -> Result<Arc<RenderDevice>>

Create a render device with no additional instanc extensions or layers.

Params

features - The physical device features required by the application.

Safety

The application is responsible for synchronizing access to all Vulkan resources and destroying the render device at exit.

pub unsafe fn create_render_device( &self, instance_extensions: &[String], instance_layers: &[String], features: PhysicalDeviceFeatures ) -> Result<Arc<RenderDevice>>

Create a render device for the application.

Params

instance_extensions - Any extensions to enable when creating the instance. Extensions for the swapchain on the current platform are added automatically and do not need to be provided.

instance_layers - Any additional layers to provide. The khronos validation layer is added automatically when debug assertions are enabled.

features - The physical device features required by the application.

Safety

The application is responsible for synchronizing access to all Vulkan resources and destroying the render device at exit.

pub unsafe fn create_vulkan_instance( &self, instance_extensions: &[String], instance_layers: &[String] ) -> Result<VulkanInstance>

Create a Vulkan instance with extensions and layers configured to such that it can present swapchain frames to the window.

Params

instance_extensions - Any extensions to enable when creating the instance. Extensions for the swapchain on the current platform are added automatically and do not need to be provided.

instance_layers - Any additional layers to provide. The khronos validation layer is added automatically when debug assertions are enabled.

Safety

The application is responsible for synchronizing access to all Vulkan resources and destroying the Vulkan instance at exit.