Module ccthw::graphics::vulkan_api

Modules

raii

Structs

BindlessTriangles

A utility for rendering high-performance textured triangles using bindless textures.

BindlessVertex

ColorPass

A utility for managing a render pass and framebuffers which target a given set of images.

Frame

The current animation Frame.

FramesInFlight

A utility for synchronizing graphics commands and submission for multiple in-flight frames.

OneTimeSubmitCommandBuffer

A utility for managing a small command pool which runs synchronous commands.

Queue

A Vulkan device queue.

RenderDevice

A combination of the VulkanInstance, LogicalDevice, and queues required by this application.

Swapchain

The Vulkan swapchain, loader, images, image views, and related data.

Texture2D

Represents a 2D rgba texture which can be used by shaders.

TextureLoader

Enums

FrameStatus

The result of a call to FramesInFlight::acquire_frame.

SwapchainStatus

Indicates that the swapchain needs a rebuild, or that the image was acquired successfully.