Module ccthw::ui

Modules

primitives

widgets

Structs

Font

This struct contains all of the information required to render rasterized glyphs on screen.

Id

This type represents a unique ID for User Interface elements.

Input

This struct holds all persistent UI input state. Things like the position of the mouse or hotkeys that are still held down.

InternalState

Internal State maintains any widget state that needs to persist between view rebuilds.

UI

The entrypoint for every UI. It manages internal state and knows how to render the UI to a frame.

Traits

UIState

Functions

id_hash

Generate a simple hash from a given string at compile time. This is used by the gen_id!() macro to generate an ID number based on the file, line, and column number of the id.

ui_screen_space_projection

Create a new projection matrix which defines the UI Screen Space based on the given width and height.