Trait ccthw::ui::widgets::CompositeWidget

pub trait CompositeWidget<IMessage, EMessage> {
    type State;
    fn id(&self) -> &Id;

    fn view(
        &mut self, 
        state: &Self::State
    ) -> Element<ComposedMessage<IMessage, EMessage>>;

    fn update(&self, state: &mut Self::State, event: IMessage) -> Result<()>;
}

Types which implemnet this type can act like a stateful collection of widges in the UI layout.

Associated Types

type State

The internal state used by this type. This is automatically retrieved from InternalStorage based on the id.

Required methods

fn id(&self) -> &Id

This instance’s Id. For predictable behavior, this should be the same every time the type is created.

fn view(
&mut self,
state: &Self::State
) -> Element<ComposedMessage<IMessage, EMessage>>

Render a view of the CompositeWidget based on the current state.

Usage

This method guaranteed to be called only once for the lifetime of the widget. Any state which needs to be preserved between calls to view should be persisted in the Self::State type.

fn update(&self, state: &mut Self::State, event: IMessage) -> Result<()>

Update this widget’s internal state based on an internal event generated by the view.