Trait ccthw::ui::widgets::Widget

pub trait Widget<Message> {
    fn handle_event(
        &mut self, 
        internal_state: &mut InternalState, 
        input: &Input, 
        event: &WindowEvent
    ) -> Result<Option<Message>>;

    fn draw_frame(
        &self, 
        internal_state: &mut InternalState, 
        frame: &mut Frame
    ) -> Result<()>;

    fn dimensions(
        &mut self, 
        internal_state: &mut InternalState, 
        max_size: &Dimensions
    ) -> Dimensions;

    fn set_top_left_position(
        &mut self, 
        internal_state: &mut InternalState, 
        position: Vec2
    );
}

Widgets are UI building blocks. Widgets ar responsible for handling system events and transforming them into an instance of their own Message when the relevant sequence of events occurs.

Layout

Widgets have two methods which are used for laying out the UI. First, the parent widget is responsible for calling [dimensions] on each of its children. The parent is then allowed to call [set_top_left_position] to position each child on the screen.

As such, the mental model is:

widgets control their own dimensions

widgets do not control their own position

parent widgets control the position of child widgets

The root widget is always positioned at (0, 0), the top left of the screen.

Required methods

fn handle_event(
&mut self,
internal_state: &mut InternalState,
input: &Input,
event: &WindowEvent
) -> Result<Option<Message>>

Handle events for this widget.

fn draw_frame(
&self,
internal_state: &mut InternalState,
frame: &mut Frame
) -> Result<()>

Render this widget to the current frame.

fn dimensions(
&mut self,
internal_state: &mut InternalState,
max_size: &Dimensions
) -> Dimensions

Compute the widget’s dimensions. A widget must not allow its dimensions to exceed the provided max size.

fn set_top_left_position(
&mut self,
internal_state: &mut InternalState,
position: Vec2
)

Set this widget’s top-left position in screen space. This is always called by the parent widget.