Module ccthw::vulkan::errors

Enums

AllocatorError

BufferError

CommandBufferError

DescriptorSetError

FenceError

FramebufferError

ImageError

InstanceError

This enum represents errors which can occur when building and handling the Vulkan instance.

PhysicalDeviceError

This enum represents the errors which can occur while attempting to find a usable physical device for the application.

PipelineError

QueueSelectionError

This enum represents errors which can occur while attempting to find all of the Vulkan command queues which are required by the application.

RenderDeviceError

This enum represents erros which can occur while working with the abstract render device.

RenderPassError

SemaphoreError

SwapchainError

VulkanDebugError

VulkanError

WindowSurfaceError