Struct ccthw::vulkan::GpuVec

pub struct GpuVec<T: Copy> {
    pub buffer: Buffer,
    /* private fields */
}

A resizable GPU Buffer which holds a contiguous slice of T’s.

Fields
buffer: Buffer

The device buffer which holds the actual data.

Implementations

impl<T: Copy> GpuVec<T>

pub fn new(
vk_dev: Arc<RenderDevice>,
vk_alloc: Arc<dyn MemoryAllocator>,
buffer_usage_flags: BufferUsageFlags,
initial_capacity: u32
) -> Result<Self, BufferError>

pub fn push_back(&mut self, value: T) -> Result<bool, BufferError>

Append a value to the data in the buffer. Grows the buffer automatically if more memory is needed.

Returns

true if the underlying buffer needed to be reallocated

false if no change was required for the underlying buffer

pub fn clear(&mut self)

Reset the length without any change to the underlying buffer.

pub fn len(&self) -> usize

The number of elements in the buffer.

pub fn len_bytes(&self) -> u64

The number of bytes in use by the buffer. e.g. self.len() * size_of<T>()