Struct ccthw::vulkan::RenderDevice

pub struct RenderDevice {
    pub physical_device: PhysicalDevice,
    pub logical_device: Device,
    pub graphics_queue: GpuQueue,
    pub present_queue: GpuQueue,
    pub swapchain: Mutex<Option<Swapchain>>,
    pub window_surface: WindowSurface,
    pub instance: Instance,
}

The render device holds the core Vulkan state and devices which are used by all parts of the application.

Fields
physical_device: PhysicalDevice

The physical device used by this application.

logical_device: Device

The Vulkan logical device used to issue commands to the physical device.

graphics_queue: GpuQueue

The GPU queue used to submit graphics commands.

present_queue: GpuQueue

The GPU queue used to submit presentation commands.

swapchain: Mutex<Option<Swapchain>>

The window’s swapchain and related resources.

window_surface: WindowSurface

The Vulkan presentation surface for the current window.

instance: Instance

The Vulkan library instance.

Implementations

impl RenderDevice

pub fn new(
instance: Instance,
window_surface: WindowSurface
) -> Result<Self, RenderDeviceError>

Create the Vulkan Render Device.

The device is initialized without a swapchain initially. An additional call to rebuild_swapchain is needed prior to any rendering operations.

pub fn name_vulkan_object<Name, Handle>(
&self,
name: Name,
object_type: ObjectType,
handle: Handle
) -> Result<(), RenderDeviceError> where
Name: Into<String>,
Handle: Handle + Copy,

Give a debug name for a vulkan object owned by this device.

Whatever name is provided here will show up in the debug logs if there are any issues detected by the validation layers.

pub fn get_supported_msaa(&self, desired: SampleCountFlags) -> SampleCountFlags

Query the device for MSAA support.

Returns

The minimum between the desired sample count and the sample count supported by the device.

e.g. if the device supports 4xMSAA and 8xMSAA is desired, this method will return 4xMSAA. Similarly, if the device supports 4xMSAA and 2xMSAA is desired, then this method will return 2xMSAA.

impl RenderDevice

pub fn with_swapchain<ReturnType, Func>(&self, func: Func) -> ReturnType where
Func: FnOnce(&Swapchain) -> ReturnType,

Perform some action with the swapchain.

Panic

If the swapchain does not exist or if the swapchain mutex cannot be acquired for any reason.

pub fn swapchain_image_count(&self) -> u32

Returns the number of images in the swapchain.

pub fn acquire_next_swapchain_image(
&self,
semaphore: Semaphore,
fence: Fence
) -> Result<usize, SwapchainError>

Acquire the next swapchain image index.

Params

semaphore - the semaphore to signal when the swapchain image is available for rendering. Can be null if uneeded.

fence - the fence to signal when the swapchain image is available for rendering. Can be null if uneeded.

pub fn rebuild_swapchain(
&self,
framebuffer_size: (u32, u32)
) -> Result<(), SwapchainError>

Rebuild the render device’s swapchain with the provided framebuffer size. Automatically handles replacing an existing swapchain if one already exists.

WARNING:

There is no internal synchronization. The application must ensure that there are no in-progress operations using the swapchain when it is replaced. This operation will block on the graphics and presentation queues draining completely, but this could be dangerous if there are out-of-order semaphore waits still pending on either queue.

e.g. if either queue contains (or could contain) a timeline semaphore wait, then make sure the corresponding signal is already queued, or else manually signal the semaphore to allow forward progress.