Module ccthw::vulkan::sync

Vulkan synchronization privite wrappers.

Structs

Fence

An owned Vulkan fence object which is automatically destroyed when dropped.

Semaphore

An owned semaphore which is automatically destroyed when it is dropped.

SemaphorePool

A semaphore pool maintains a collection of binary semaphores which are available for re-use. Unused semaphores are automatically destroyed when the pool is dropped.

Enums

FenceError

SemaphoreError