Trait ccthw::vulkan::ComposableAllocator

pub trait ComposableAllocator {
    unsafe fn allocate(
        &mut self, 
        allocate_info: MemoryAllocateInfo, 
        alignment: u64
    ) -> Result<Allocation, AllocatorError>;

    unsafe fn free(
        &mut self, 
        allocation: &Allocation
    ) -> Result<(), AllocatorError>;
}

The device memory allocation interface. This is the compositional API for GPU memory allocation.

Required methods

unsafe fn allocate(
&mut self,
allocate_info: MemoryAllocateInfo,
alignment: u64
) -> Result<Allocation, AllocatorError>

Allocate device memory with the provided type index and size.

unsafe because

it is the responsibility of the caller to free the returned memory when it is no longer in use

implementations do not generally check that the memory type index in allocate_info is the correct memory type index, the arguments are assumed to be correct

unsafe fn free(&mut self, allocation: &Allocation) -> Result<(), AllocatorError>

Free an allocated piece of device memory.

unsafe because

it is the responsibility of the caller to know when the GPU is no longer using the allocation