Trait ccthw::vulkan::MemoryAllocator

pub trait MemoryAllocator {
    unsafe fn allocate_memory(
        &self, 
        memory_requirements: MemoryRequirements, 
        property_flags: MemoryPropertyFlags
    ) -> Result<Allocation, AllocatorError>;

    unsafe fn free(&self, allocation: &Allocation) -> Result<(), AllocatorError>;
}

Required methods

unsafe fn allocate_memory(
&self,
memory_requirements: MemoryRequirements,
property_flags: MemoryPropertyFlags
) -> Result<Allocation, AllocatorError>

Allocate GPU memory based on a given set of requirements.

unsafe

it is the responsibility of the caller to free the returned memory when it is no longer in use

unsafe fn free(&self, allocation: &Allocation) -> Result<(), AllocatorError>

Free an allocated piece of device memory.

unsafe because

it is the responsibility of the caller to know when the GPU is no longer using the allocation